requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
click==8.2.1
click-plugins==1.1.1.2
daphne==4.2.1
Django==5.2.5
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-filter==25.1
django-redis==6.0.0
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
kombu==5.5.4
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==6.4.0
setuptools<81
sqlparse==0.5.3
vine==5.1.0
gunicorn==21.2.0
drf-yasg==1.21.6
drf-spectacular==0.28.0
pytest
pytest-django